+++ ndk/src/events/kinds/dvm/request.ts
@@ -1,47 +1,30 @@
 import type { NDK } from "../../../ndk/index.js";
+import { NDKSigner } from "../../../signers/index.js";
+import { NDKUser } from "../../../user/index.js";
 import type { NDKTag, NostrEvent } from "../../index.js";
 import { NDKEvent } from "../../index.js";
-import type { NDKDvmJobFeedbackStatus } from "./NDKDVMJobFeedback.js";
-import { NDKDVMJobFeedback } from "./NDKDVMJobFeedback.js";
-import { NDKDVMJobResult } from "./NDKDVMJobResult.js";
+// import type { NDKDvmJobFeedbackStatus } from "./NDKDVMJobFeedback.js";
+// import { NDKDVMJobFeedback } from "./NDKDVMJobFeedback.js";
+// import { NDKDVMJobResult } from "./NDKDVMJobResult.js";
 
 /**
- * NIP-90: Data vending machine
+ * NIP-90: Data vending machine request
  *
  * A generic Job request class for Data Vending Machines
+ *
+ * @example
+ * const request = new NDKDVMRequest(ndk);
+ * request.kind = NDKKind.DVMReqTextExtraction;
+ * request.addInput(["https://allenfarrington.medium.com/modeling-bitcoin-value-with-vibes-99eca0997c5f", "url"])
+ * await request.publish()
  */
 export class NDKDVMRequest extends NDKEvent {
     constructor(ndk: NDK | undefined, event?: NostrEvent) {
         super(ndk, event);
-
-        if (ndk) {
-            this.tags.push(["relays", ...ndk.pool.urls()]);
-        }
     }
 
     static from(event: NDKEvent) {
         return new NDKDVMRequest(event.ndk, event.rawEvent());
-    }
-
-    /**
-     * Create a new job feedback for this request
-     */
-    public createFeedback(status: NDKDvmJobFeedbackStatus | string) {
-        const feedback = new NDKDVMJobFeedback(this.ndk);
-        feedback.status = status;
-        feedback.tag(this, "job");
-
-        return feedback;
-    }
-
-    /**
-     * Create a new result event for this request
-     */
-    public createResult(data?: NostrEvent) {
-        const event = new NDKDVMJobResult(this.ndk, data);
-        event.jobRequest = this;
-
-        return event;
     }
 
     set bid(msatAmount: number | undefined) {
@@ -97,5 +80,33 @@ export class NDKDVMRequest extends NDKEvent {
     public getParam(name: string): string | undefined {
         const paramTag = this.getMatchingTags("param").find((t: NDKTag) => t[1] === name);
         return paramTag ? paramTag[2] : undefined;
+    }
+
+    /**
+     * Enables job encryption for this event
+     * @param dvm DVM that will receive the event
+     * @param signer Signer to use for encryption
+     */
+    public async encryption(dvm: NDKUser, signer?: NDKSigner) {
+        const dvmTags = ["i", "param", "output", "relays", "bid"];
+        const tags = this.tags.filter((t) => dvmTags.includes(t[0]));
+
+        // remove all tags that will be encrypted
+        this.tags = this.tags.filter((t) => !dvmTags.includes(t[0]));
+        this.content = JSON.stringify(tags);
+        this.tag(dvm);
+        this.tags.push(["encrypted"]);
+
+        // encrypt the event
+        await this.encrypt(dvm, signer);
+    }
+
+    /**
+     * Sets the DVM that will receive the event
+     */
+    set dvm(dvm: NDKUser | undefined) {
+        this.removeTag("p");
+
+        if (dvm) this.tag(dvm);
     }
 }