+++ ndk/src/index.ts
@@ -1,17 +1,24 @@
 export * from "./cache/index.js";
 export * from "./user/index.js";
+export * from "./user/pin.js";
 export * from "./events/index.js";
 
 // Kinds
 export * from "./events/kinds/index.js";
 export * from "./events/kinds/article.js";
+export * from "./events/kinds/classified.js";
 export * from "./events/kinds/video.js";
 export * from "./events/kinds/highlight.js";
 export * from "./events/kinds/NDKRelayList.js";
-export * from "./events/kinds/dvm/index.js";
 export * from "./events/kinds/lists/index.js";
 export * from "./events/kinds/repost.js";
 export * from "./events/kinds/nip89/NDKAppHandler.js";
+export * from "./events/kinds/subscriptions/tier.js";
+export * from "./events/kinds/subscriptions/amount.js";
+export * from "./events/kinds/subscriptions/subscription-start.js";
+export * from "./events/kinds/subscriptions/receipt.js";
+export * from "./events/kinds/dvm/index.js";
+export * from "./nwc/index.js";
 
 export * from "./events/kinds/simple-group/index.js";
 
@@ -27,6 +34,8 @@ export * from "./signers/private-key/index.js";
 export * from "./subscription/index.js";
 export * from "./subscription/grouping.js";
 export * from "./user/profile.js";
+
+export * from "./dvm/schedule.js";
 
 export { NDK as default, NDKConstructorParams } from "./ndk/index.js";
 export { NDKZapInvoice, zapInvoiceFromEvent } from "./zap/invoice.js";