+++ ndk/src/relay/connectivity.ts
@@ -3,11 +3,13 @@ import { relayInit } from "nostr-tools";
 
 import type { NDKRelay, NDKRelayConnectionStats } from ".";
 import { NDKRelayStatus } from ".";
+import { runWithTimeout } from "../utils/timeout";
 
 export class NDKRelayConnectivity {
     private ndkRelay: NDKRelay;
     private _status: NDKRelayStatus;
     public relay: Relay;
+    private timeoutMs?: number;
     private connectedAt?: number;
     private _connectionStats: NDKRelayConnectionStats = {
         attempts: 0,
@@ -15,6 +17,7 @@ export class NDKRelayConnectivity {
         durations: [],
     };
     private debug: debug.Debugger;
+    private reconnectTimeout: any;
 
     constructor(ndkRelay: NDKRelay) {
         this.ndkRelay = ndkRelay;
@@ -25,33 +28,21 @@ export class NDKRelayConnectivity {
         this.relay.on("notice", (notice: string) => this.handleNotice(notice));
     }
 
-    async initiateAuth(filter = { limit: 1 }): Promise<void> {
-        this.debug("Initiating authentication");
-        const authSub = this.relay.sub([filter], { id: "auth-test" });
-        authSub.on("eose", () => {
-            // we didn't need to authenticate
-            authSub.unsub();
-            this._status = NDKRelayStatus.CONNECTED;
-            this.ndkRelay.emit("ready");
-            this.debug("Authentication not required");
-            authSub.unsub();
-        });
-        this.debug("Authentication request started");
-    }
+    public async connect(timeoutMs?: number): Promise<void> {
+        if (this.reconnectTimeout) {
+            clearTimeout(this.reconnectTimeout);
+            this.reconnectTimeout = undefined;
+        }
 
-    public async connect(): Promise<void> {
+        timeoutMs ??= this.timeoutMs;
+        if (!this.timeoutMs && timeoutMs) this.timeoutMs = timeoutMs;
+
         const connectHandler = () => {
             this.updateConnectionStats.connected();
 
-            if (!this.ndkRelay.authRequired) {
-                this._status = NDKRelayStatus.CONNECTED;
-                this.ndkRelay.emit("connect");
-                this.ndkRelay.emit("ready");
-            } else {
-                this._status = NDKRelayStatus.AUTH_REQUIRED;
-                this.ndkRelay.emit("connect");
-                this.initiateAuth();
-            }
+            this._status = NDKRelayStatus.CONNECTED;
+            this.ndkRelay.emit("connect");
+            this.ndkRelay.emit("ready");
         };
 
         const disconnectHandler = () => {
@@ -77,7 +68,7 @@ export class NDKRelayConnectivity {
                     if (this._status === NDKRelayStatus.AUTHENTICATING) {
                         this.debug("Authentication policy finished");
                         this._status = NDKRelayStatus.CONNECTED;
-                        this.ndkRelay.emit("ready");
+                        this.ndkRelay.emit("authed");
                     }
                 }
             } else {
@@ -87,7 +78,10 @@ export class NDKRelayConnectivity {
 
         try {
             this.updateConnectionStats.attempt();
-            this._status = NDKRelayStatus.CONNECTING;
+            if (this._status === NDKRelayStatus.DISCONNECTED)
+                this._status = NDKRelayStatus.CONNECTING;
+            else
+                this._status = NDKRelayStatus.RECONNECTING;
 
             this.relay.off("connect", connectHandler);
             this.relay.off("disconnect", disconnectHandler);
@@ -95,10 +89,15 @@ export class NDKRelayConnectivity {
             this.relay.on("disconnect", disconnectHandler);
             this.relay.on("auth", authHandler);
 
-            await this.relay.connect();
+            await runWithTimeout(
+                this.relay.connect,
+                timeoutMs,
+                "Timed out while connecting"
+            );
         } catch (e) {
             this.debug("Failed to connect", e);
             this._status = NDKRelayStatus.DISCONNECTED;
+            this.handleReconnection();
             throw e;
         }
     }
@@ -150,13 +149,16 @@ export class NDKRelayConnectivity {
             // }, 60000);
         }
 
-        this.ndkRelay.emit("notice", this, notice);
+        this.ndkRelay.emit("notice", this.relay, notice);
     }
 
     /**
      * Called when the relay is unexpectedly disconnected.
      */
     private handleReconnection(attempt = 0): void {
+        if (this.reconnectTimeout) return;
+        this.debug("Attempting to reconnect", { attempt });
+
         if (this.isFlapping()) {
             this.ndkRelay.emit("flapping", this, this._connectionStats);
             this._status = NDKRelayStatus.FLAPPING;
@@ -165,10 +167,12 @@ export class NDKRelayConnectivity {
 
         const reconnectDelay = this.connectedAt
             ? Math.max(0, 60000 - (Date.now() - this.connectedAt))
-            : 0;
+            : 5000 * (this._connectionStats.attempts+1);
 
-        setTimeout(() => {
+        this.reconnectTimeout = setTimeout(() => {
+            this.reconnectTimeout = undefined;
             this._status = NDKRelayStatus.RECONNECTING;
+            this.debug(`Reconnection attempt #${attempt}`)
             this.connect()
                 .then(() => {
                     this.debug("Reconnected");
@@ -176,15 +180,18 @@ export class NDKRelayConnectivity {
                 .catch((err) => {
                     this.debug("Reconnect failed", err);
 
-                    if (attempt < 5) {
+                    if (attempt < 10) {
                         setTimeout(() => {
                             this.handleReconnection(attempt + 1);
-                        }, 60000);
+                        }, 1000 * (attempt+1) ^ 2);
                     } else {
-                        this.debug("Reconnect failed after 5 attempts");
+                        this.debug("Reconnect failed after 10 attempts");
                     }
                 });
         }, reconnectDelay);
+
+        this.debug("Reconnecting in", reconnectDelay);
+        this._connectionStats.nextReconnectAt = Date.now() + reconnectDelay;
     }
 
     /**