+++ ndk-svelte-components/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @nostr-dev-kit/ndk-svelte-components
 
+## 2.2.8
+
+### Patch Changes
+
+-   Updated dependencies
+-   Updated dependencies
+    -   @nostr-dev-kit/ndk@2.5.1
+
+## 2.2.7
+
+### Patch Changes
+
+-   Updated dependencies [e08fc74]
+    -   @nostr-dev-kit/ndk@2.5.0
+
+## 2.2.6
+
+### Patch Changes
+
+-   55011e3: fix rendering issues on RelayList
+-   Updated dependencies [111c1ea]
+-   Updated dependencies [5c0ae51]
+-   Updated dependencies [6f5ea49]
+-   Updated dependencies [3738d39]
+-   Updated dependencies [d22239a]
+    -   @nostr-dev-kit/ndk@2.4.1
+
+## 2.2.5
+
+### Patch Changes
+
+-   Updated dependencies [b9bbf1d]
+    -   @nostr-dev-kit/ndk@2.4.0
+
 ## 2.2.4
 
 ### Patch Changes